apply transpilation result to hoisted packages
This commit fixes the bug when only nested node_modules
deployments would receive linkage updates, not the versions
of the package stored in the root node_modules. This bug
occurred, beceause of invalid pattern expression in the
command responsible for finding destinations for where
the program must apply transpilation result.