link in root of workspace
This commit makes so that when Oink is called
in one of the child packages in a workspace,
root node_modules would be prioritized during during
linking. This is because, in yarn workspaces, hoisting
makes so that all semver compatible packages appear
in node_modules folder in root of the workspace, not
in individual child packages. PNPM operates in a similar
way, but has symlinks in node_modules. Logic in this commit
doesn't care, and always fallbacks to root of workspace,
since it's the safest option. From there, Oink will
find all nested node_modules references anyways.