fix: introduce middleware for imports
This commit introduces custom transpile.js script,
which pipes into esbuild and is now used in transpilation pipeline.
It removes new format of imports that may be used in underlying
dependencies, so that imports like, for example, "node:fs" become "fs".
This allows for the Oink binary to be used on bigger variety of node
versions and setups.