chore: break out command handlers from cli submodule
This commit restructures the codebase of the application
by separating commander program set up and handlers for
commands like "learn", "link" and "forget". Now a new submodule
named "commands" exists, which must only contain raw handlers
for the commands. This allows for the command handlers to be
tested in an isolated environment. The "cli" submodule is now
depended on "commands" submodule, as it imports handlers and
injects them into commander program that gets created on application
init.