chore: reset core tests
This commit resets all existing unit test cases for the "core"
submodule. Previously, the tests were ran while heavily relying
on stubbing and mocking. This has proven to slow down development
and NOT properly ensure that the application is bug-free. This
commit is the first step for the tests reworking with "memfs".