fix: cleanup nested folders in forget
This commit fixes the issue when packages in
spaces (like "@svelte/react") would leave traces
in the links registry folder (in ".config/") when
executing forget command.

Running forget while having "@svelte/react" remembered,
would leave folder "@svelte" in the registry, with no symlink
inside. This technically didn't lead to any logical issues, but
created junk folders that would never get cleaned up.

This commit fixes the issue by introducing an additional check
that runs whenever a link is deleted to verify that parent folder
is not empty. This makes so that any nested folder is deleted
when it becomes empty, but also when links registry folder becomes
empty it would get deleted as well.

This change allows to maintain clean file structure.