fix(docs): add base url to links
This commit introduces a utility hrefWithBaseURL,
which allows to append baseURL to any href link. Previously,
the links were broken. If they were relative, they'd navigate
to a "sub" of the current page (which of course would usually lead to
anything). If they were absolute, they'd ignore the baseURL and navigate
to a route starting with "/" (root) instead of specified baseURL.

This commit fixes the issue by introducing and applying utility
that normalizes any relative paths to make them absolute by appending
BASE_URL from env.